// 10-Professional_React/products-explorer-with-cart-improved/src/types/product.ts
export type Review = {
  rating: number;
  comment: string;
  date: string;
  reviewerName: string;
}

export type Product = {
  id: number,
  thumbnail: string;
  title: string;
  category: string;
  price: number;
  rating: number;
  availabilityStatus: string;
  stock: number;
  warrantyInformation: string;
  shippingInformation: string;
  reviews: Review[]
}